Fixed 32/64 architecture bug
git-svn-id: https://bne-fs-pf-001.euclideon.local/svn/Dev/trunk/src@157 f54097bf-93f5-7042-b87d-fd266665e7ff

+++ udPlatform/Source/udFileHandler_HTTP.cpp
@@ -329,7 +329,7 @@ static udResult udFileHandler_HTTPSeekRead(udFile *pBaseFile, void *pBuffer, siz
 {
   udResult result;
   udFile_HTTP *pFile = static_cast<udFile_HTTP *>(pBaseFile);
-  size_t offset = pFile->currentOffset;
+  int64_t offset = pFile->currentOffset;
 
   switch (seekWhence)
   {
